rehashed activation from designation to membership
A lot of changes introduced with this commit. Decorators via draper gem was used to
present views. And activation of designations was discarded in favour of activation
of team membership. As team ownership is recorded, assigning the members to the
owner was easily done. Plenty of methods were transfered from the models to their
respective decorators in order to shift the logic out of the view files.